+++ src/console/controllers/FieldsController.php
@@ -4,10 +4,12 @@ namespace benf\neo\console\controllers;
 
 use benf\neo\elements\Block;
 use benf\neo\Field;
+use benf\neo\Plugin as Neo;
 use Craft;
 use craft\console\Controller;
 use craft\db\Query;
 use craft\db\Table;
+use craft\enums\PropagationMethod;
 use craft\helpers\Queue;
 use craft\i18n\Translation;
 use craft\queue\jobs\ApplyNewPropagationMethod;
@@ -71,10 +73,10 @@ class FieldsController extends Controller
         $withPropagationMethod = $this->withPropagationMethod
             ? explode(',', $this->withPropagationMethod)
             : [
-                Field::PROPAGATION_METHOD_NONE,
-                Field::PROPAGATION_METHOD_SITE_GROUP,
-                Field::PROPAGATION_METHOD_LANGUAGE,
-                Field::PROPAGATION_METHOD_CUSTOM,
+                PropagationMethod::Custom->value,
+                PropagationMethod::Language->value,
+                PropagationMethod::None->value,
+                PropagationMethod::SiteGroup->value,
             ];
 
         // If not reapplying by block structure, just do one for every field
@@ -89,7 +91,7 @@ class FieldsController extends Controller
 
             $neoFields = array_filter($fieldsService->getAllFields(), function($field) use ($setIds, $withPropagationMethod) {
                 return $field instanceof Field &&
-                    in_array($field->propagationMethod, $withPropagationMethod) &&
+                    in_array($field->propagationMethod->value, $withPropagationMethod) &&
                     (empty($setIds) || isset($setIds[$field->id]));
             });
 
@@ -128,7 +130,7 @@ class FieldsController extends Controller
                     continue;
                 }
 
-                $fieldPropagationMethod[$fieldId] = $field->propagationMethod;
+                $fieldPropagationMethod[$fieldId] = $field->propagationMethod->value;
             }
 
             if (in_array($fieldPropagationMethod[$fieldId], $withPropagationMethod)) {